+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2017-10-23
+----------
+
+ * allow enumerations that have "null" values; this defers null handling into the Constructor
+
 2017-09-02
 ----------